Genre cleanup: removed use of AutoMapper. Moved mapping from controller to service with new manual mapping method

+++ MovieDatabase.Core/Services/GenreService.cs
@@ -1,4 +1,7 @@
-﻿using MovieDatabase.Core.Models;
+﻿using FluentValidation;
+using MovieDatabase.Core.Mapping;
+using MovieDatabase.Core.Models;
+using MovieDatabase.Core.Models.Dto;
 using MovieDatabase.Core.Repository.IRepository;
 using System.Linq.Expressions;
 
@@ -7,37 +10,59 @@ namespace MovieDatabase.Core.Services
     public class GenreService : IGenreService
     {
         private readonly IGenreRepository _genreRepository;
+        private readonly IValidator<Genre> _genreValidator;
 
-        public GenreService(IGenreRepository genreRepository)
+        public GenreService(IGenreRepository genreRepository, IValidator<Genre> genreValidator)
         {
             _genreRepository = genreRepository;
+            _genreValidator = genreValidator;
         }
 
-        public async Task<IEnumerable<Genre>> GetAllAsync(Expression<Func<Genre, bool>>? filter = null, string? includeProperties = null)
+        public async Task<IEnumerable<GenreDto>> GetAllAsync(Expression<Func<Genre, bool>>? filter = null, string? includeProperties = null)
         {
-            return await _genreRepository.GetAllAsync(filter, includeProperties);
+            var genres = await _genreRepository.GetAllAsync(filter, includeProperties);
+            return genres.Select(g => g.MapToGenreResponse());
         }
 
-        public async Task<Genre?> GetByIdAsync(Expression<Func<Genre, bool>>? filter = null, bool tracked = true, string? includeProperties = null)
+        public async Task<GenreDto?> GetByIdAsync(int id, bool tracked = true, string? includeProperties = null)
         {
-            return await _genreRepository.GetByIdAsync(filter, tracked, includeProperties);
+            var genre = await _genreRepository.GetByIdAsync(id, tracked, includeProperties);
+            
+            if (genre == null)
+                return null;
+
+            return genre.MapToGenreResponse();
         }
-        public async Task<Genre> CreateAsync(Genre newGenre)
+        public async Task<GenreDto> CreateAsync(GenreCreateDto newGenre)
         {
-            await _genreRepository.CreateAsync(newGenre);
-            return newGenre;
+            var genre = newGenre.MapToGenre();
+            await _genreValidator.ValidateAndThrowAsync(genre);
+            await _genreRepository.CreateAsync(genre);
+            return genre.MapToGenreResponse();
         }
 
-        public async Task DeleteAsync(Genre genre)
+        public async Task<bool> DeleteAsync(int id)
         {
-            await _genreRepository.RemoveAsync(genre);
-            return;
+            var genre = await _genreRepository.GetByIdAsync(id, tracked: true);
+            
+            if (genre == null)
+                return false;
+
+            return await _genreRepository.RemoveAsync(genre);
         }
 
-        public async Task<Genre> UpdateAsync(Genre updatedGenre)
+        public async Task<GenreDto?> UpdateAsync(GenreUpdateDto updatedGenre)
         {
-            await _genreRepository.UpdateAsync(updatedGenre);
-            return updatedGenre;
+            var existingGenre = await _genreRepository.GetByIdAsync(updatedGenre.Id, tracked: true);
+
+            if (existingGenre == null)
+                return null;
+
+            existingGenre = updatedGenre.MapToGenre(existingGenre);
+            await _genreValidator.ValidateAndThrowAsync(existingGenre);
+            await _genreRepository.UpdateAsync(existingGenre);
+
+            return existingGenre.MapToGenreResponse();
         }
 
         public async Task<List<Genre>?> GetGenresByNamesAsync(List<string> genreNames)
